+++ src/main/java/entgraClothingapp/application/dto/request/CreateItemDto.java
@@ -3,7 +3,7 @@ package entgraClothingapp.application.dto.request;
 import lombok.Data;
 @Data
 public class CreateItemDto {
-        private int id;
+        private long id;
         private String code;
         private String itemTitle;
         private String itemType;
@@ -15,9 +15,9 @@ public class CreateItemDto {
         private String status;
         private int numberOfItems;
         private int buyingPrice;
-        private int startingPrice;
-        private int profitPercentage;
-        private int salePercentage;
-        private int salePrice;
+        private float startingPrice;
+        private float profitPercentage;
+        private float salePercentage;
+        private float salePrice;
         private int stockClearingPrice;
 }
